% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edigs-package.R
\docType{package}
\name{edigs-package}
\alias{edigs}
\alias{edigs-package}
\title{edigs: Efficiently Create Economic Diagrams}
\description{
A system for 'declaratively' creating economic diagrams, consisting of a suite of primitive and high-level functions for drawing drawing common economic diagrams and adding elements (e.g. labelled curves, labelled points, etc.) to them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lyuben-ivanov/edigs}
  \item Report bugs at \url{https://github.com/lyuben-ivanov/edigs/issues}
}

}
\author{
\strong{Maintainer}: Lyuben Ivanov \email{l.ivanov@feb.uni-sofia.bg} (\href{https://orcid.org/0000-0003-3900-8897}{ORCID})

}
\keyword{internal}
